Add `CartesianGestureHandler`
separates it out of the body of `CartesianChart`
